Extract QuestionController interface.
