+++ app-ktor/src/main/kotlin/ru/otus/opinion/backend/ktor/plugins/Routing.kt
@@ -4,13 +4,14 @@ import io.ktor.application.*
 import io.ktor.response.*
 import io.ktor.routing.*
 import ru.otus.opinion.backend.ktor.controllers.QuestionController
+import ru.otus.opinion.backend.ktor.controllers.QuestionControllerImpl
 import ru.otus.opinion.bakend.services.QuestionService
 import ru.otus.opinion.bakend.services.QuestionServiceImpl
 
 fun Application.configureRouting() {
 
     val questionService: QuestionService = QuestionServiceImpl()
-    val questionController = QuestionController(questionService)
+    val questionController: QuestionController = QuestionControllerImpl(questionService)
 
     routing {
         get("/") {
